requirements.txt
openai==0.27.2
Pillow==9.4.0